Make testing graceful
As testing becomes more complex, it needs to be possible to spin
up a SUT and have it torn down at the end of the test.

This PR is the first step in that direction, using exechelper.WithGracePeriod
to kill the SUT with SIGTERM, to allow it to kill of its child vpp and vppagent
processes.

Signed-off-by: Ed Warnicke <hagbard@gmail.com>